+++ go.mod
@@ -4,5 +4,5 @@ go 1.12
 
 require (
 	github.com/gorilla/websocket v1.4.1
-	gopkg.in/yaml.v2 v2.2.4
+	gopkg.in/yaml.v2 v2.2.8
 )
