+++ src/lib/handlers/handle_jitter_start.go
@@ -7,8 +7,8 @@ import (
 	"github.com/zairza-cetb/bench-routes/src/lib/filters"
 	"github.com/zairza-cetb/bench-routes/src/lib/modules/jitter"
 	"github.com/zairza-cetb/bench-routes/src/lib/utils"
-	"github.com/zairza-cetb/bench-routes/src/lib/utils/logger"
-	"github.com/zairza-cetb/bench-routes/src/lib/utils/parser"
+	"github.com/zairza-cetb/bench-routes/src/lib/logger"
+	"github.com/zairza-cetb/bench-routes/src/lib/parser"
 )
 
 // HandleJitterStart handle the route "start"
