+++ src/lib/modules/ping/ping_test.go
@@ -5,7 +5,7 @@ import (
 	"testing"
 
 	"github.com/zairza-cetb/bench-routes/src/lib/utils"
-	"github.com/zairza-cetb/bench-routes/src/lib/utils/parser"
+	"github.com/zairza-cetb/bench-routes/src/lib/parser"
 	"github.com/zairza-cetb/bench-routes/tsdb"
 
 	"github.com/zairza-cetb/bench-routes/src/lib/filters"
